refactor: separation
