+++ dictknife/swaggerknife/stream/names.py
@@ -1,0 +1,4 @@
+class schema:
+    primitive = "primitiveSchema"
+    object = "objectSchema"
+    array = "arraySchema"
